requirements.txt
streamlit
pandas
openpyxl
matplotlib
reportlab
Pillow